requirements.txt
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.0
asgi-lifespan==2.1.0
asyncpg==0.27.0
certifi==2023.5.7
cfgv==3.3.1
click==8.1.3
databases==0.6.2
distlib==0.3.6
factory-boy==3.2.1
Faker==18.11.1
fastapi==0.98.0
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
Mako==1.2.4
MarkupSafe==2.1.3
nodeenv==1.8.0
ormar==0.12.2
packaging==23.1
platformdirs==3.7.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2-binary==2.9.6
pydantic==1.10.8
pytest==7.4.0
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
uvloop==0.17.0
virtualenv==20.23.1
watchfiles==0.19.0
websockets==11.0.3
